requirements.txt
pandas
numpy
openpyxl
odfpy                
pyarrow
xmltodict            
ruamel.yaml          
openai
transformers         
sentence-transformers 
langchain            
jupyterlab
ipykernel
pytest
black
isort
mypy
pre-commit
